Removed useless class
